removing auth_id from User

+++ movie-api/src/main/java/ch/heigvd/amt/bidirhandshake/movieapi/repositories/UserRepository.java
@@ -5,8 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 
 public interface UserRepository extends CrudRepository<User, Long> {
-
     User findByUsername(String username);
 
-    User findById(long id);
+    User findById(Integer id);
 }
